+++ kibot/out_compress.py
@@ -14,7 +14,8 @@ from tarfile import open as tar_open
 from collections import OrderedDict
 from .gs import GS
 from .kiplot import config_output, get_output_dir, run_output
-from .misc import MISSING_TOOL, WRONG_INSTALL, W_EMPTYZIP, WRONG_ARGUMENTS, INTERNAL_ERROR, ToolDependency, ToolDependencyRole
+from .misc import (MISSING_TOOL, WRONG_INSTALL, W_EMPTYZIP, WRONG_ARGUMENTS, INTERNAL_ERROR, ToolDependency,
+                   ToolDependencyRole, TRY_INSTALL_CHECK)
 from .optionable import Optionable, BaseOptions
 from .registrable import RegOutput, RegDependency
 from .macros import macros, document, output_class  # noqa: F401
@@ -22,7 +23,7 @@ from . import log
 
 logger = log.get_logger()
 RegDependency.register(ToolDependency('compress', 'RAR', 'https://www.rarlab.com/',
-                                      url_down='https://www.rarlab.com/download.htm',
+                                      url_down='https://www.rarlab.com/download.htm', help_option='-?',
                                       roles=ToolDependencyRole(desc='Compress in RAR format')))
 
 
@@ -107,6 +108,7 @@ class CompressOptions(BaseOptions):
                 check_output(cmd, stderr=STDOUT)
             except FileNotFoundError:
                 logger.error('Missing `rar` command, install it')
+                logger.error(TRY_INSTALL_CHECK)
                 exit(MISSING_TOOL)
             except CalledProcessError as e:
                 logger.error('Failed to invoke rar command, error {}'.format(e.returncode))