+++ kibot/registrable.py
@@ -170,8 +170,8 @@ class RegDependency(Registrable):
 
 
 # Here we register some global dependencies
-RegDependency.register(ToolDependency('global', 'colorama', is_python=True,
+RegDependency.register(ToolDependency('global', 'Colorama', is_python=True,
                        roles=ToolDependencyRole(desc='get color messages in a portable way')))
-RegDependency.register(ToolDependency('global', 'distutils', is_python=True))
-RegDependency.register(ToolDependency('global', 'requests', is_python=True))
-RegDependency.register(ToolDependency('global', 'PyYAML', is_python=True, deb='python3-yaml'))
+RegDependency.register(ToolDependency('global', 'Distutils', is_python=True))
+RegDependency.register(ToolDependency('global', 'Requests', is_python=True))
+RegDependency.register(ToolDependency('global', 'PyYAML', is_python=True, deb='python3-yaml', module_name='yaml'))